refactor delegate subagent resolver split
